requirement.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.4
idna==2.6
psycopg2==2.7.3.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
tornado==4.5.2
tornado-redis==2.4.18
urllib3==1.22
